requirements.txt
alembic==0.9.5
aniso8601==1.2.1
astroid==1.5.3
autoenv==1.0.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
codeclimate-test-reporter==0.2.3
colorama==0.3.9
coverage==4.0.3
Flask==0.12.2
Flask-API==0.7.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.0
flask-paginate==0.5.0
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pbr==3.1.1
pep8==1.7.0
psycopg2==2.7.3
py==1.4.34
pycparser==2.18
PyJWT==1.5.3
pylint==1.7.2
pytest==3.2.1
pytest-cov==2.5.1
python-coveralls==2.9.1
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
validate-email==1.3
SQLAlchemy==1.1.13
stevedore==1.25.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.12.2
wrapt==1.10.11
